+++ src/operations.cc
@@ -191,12 +191,20 @@ namespace sharp {
       VImage alpha = image[image.bands() - 1];
       return RemoveAlpha(image)
         .colourspace(VIPS_INTERPRETATION_LCH)
-        .linear({brightness, saturation, 1}, {0, 0, static_cast<double>(hue)})
+        .linear(
+          { brightness, saturation, 1},
+          { 0.0, 0.0, static_cast<double>(hue) }
+        )
+        .colourspace(VIPS_INTERPRETATION_sRGB)
         .bandjoin(alpha);
     } else {
       return image
         .colourspace(VIPS_INTERPRETATION_LCH)
-        .linear({brightness, saturation, 1}, {0, 0, static_cast<double>(hue)});
+        .linear(
+          { brightness, saturation, 1 },
+          { 0.0, 0.0, static_cast<double>(hue) }
+        )
+        .colourspace(VIPS_INTERPRETATION_sRGB);
     }
   }