+++ docs/changelog.md
@@ -14,6 +14,9 @@ Requires libvips v8.8.1.
   [#1952](https://github.com/lovell/sharp/pull/1952)
   [@pouya-eghbali](https://github.com/pouya-eghbali)
 
+* Ensure `modulate` and other colour-based operations can co-exist.
+  [#1958](https://github.com/lovell/sharp/issues/1958)
+
 #### v0.23.2 - 28<sup>th</sup> October 2019
 
 * Add `background` option to tile output operation.